Minor corrections

+++ kcpredict/data/make_data.py
@@ -80,7 +80,12 @@ def main(input_file, output_file, visualize=True):
     # Optionally visualize the data
     if visualize:
         data.plot(subplots=True, figsize=(10, 16))
-        plt.savefig(ROOT_DIR / "visualization/data" / "raw_data.png")
+        savepath = ROOT_DIR / "visualization/data/raw_data.png"
+        if savepath.exists():
+            plt.savefig(ROOT_DIR / "visualization/data" / "raw_data.png")
+        else:
+            plt.show()
+            logging.info("Visualization folder not found. Skipping visualization.")
 
     logging.info(f'\n\n{"-"*21}')
     return None
